Object Visuals
This is meant to be used on any object you want to have a visual on for the Object Interaction script. This is meant to make the object move around smoother. Attach this script to the game object that has the mesh renderer on it and a transform. Then drag in the object it is meant to follow.